+++ guides/CHANGELOG.md
@@ -1,1 +1,5 @@
+*   Add "back to top" button to guides
+
+    *Alysson Rosa*
+
 Please check [7-0-stable](https://github.com/rails/rails/blob/7-0-stable/guides/CHANGELOG.md) for previous changes.